refactor(apps/web): restructure album create into create/ subfolder
Made-with: Cursor
